Update: 사용자 관련 기능 수정

+++ src/domain/usecase/user/RemoveUserUseCase.ts
@@ -1,6 +1,8 @@
 import { IUseCase } from '@/common/usecase/IUseCase';
 import { IUserRepository } from '../../repository/IUserRepository';
 import { IRemoveUserPort } from '@/domain/port/user/IRemoveUserPort';
+import { Exception } from '@/common/exception/Exception';
+import { Code } from '@/common/code/Code';
 
 export class RemoveUserUseCase implements IUseCase<IRemoveUserPort, void> {
   constructor(private readonly userRepository: IUserRepository) {}
@@ -10,8 +12,11 @@ export class RemoveUserUseCase implements IUseCase<IRemoveUserPort, void> {
    * @param port IRemoveUserPort
    * @step_1 port로 받아온 id값으로 유저를 삭제한다.
    * @returns void
+   * todo: 인증 과정이 필요
    */
   async execute(port?: IRemoveUserPort): Promise<void> {
+    const userExist = await this.userRepository.getUserById(port.id);
+    if (!userExist) throw Exception.new({ code: Code.NOT_FOUND, overrideMessage: '사용자 없음' });
     await this.userRepository.removeUserById(port.id);
   }
 }
